add some healthy meals

+++ database/migrations/2023_08_28_134046_create_suggested_meals_table.php
@@ -15,10 +15,10 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->text('details');
-            $table->integer('calories');
-            $table->integer('protein');
-            $table->integer('fats');
-            $table->integer('carbs');
+            $table->double('calories');
+            $table->double('protein');
+            $table->double('fats');
+            $table->double('carbs');
             $table->string('image');
             $table->timestamps();
         });
